pyproject.toml
[project]
name = "mcp-tool-servers"
version = "0.1.0"
description = "Shared MCP tool servers for the agent marketplace"
authors = [
    {name = "None"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastmcp (>=3.0.0)",
    "tavily-python (>=0.7.23,<0.8.0)",
    "firecrawl-py (>=4.19.0,<5.0.0)",
    "yfinance (>=1.2.0,<2.0.0)",
    "arxiv (>=2.4.1,<3.0.0)",
    "pymupdf (>=1.27.2,<2.0.0)",
    "pymupdf4llm (>=1.27.2.1,<2.0.0.0)",
    "pinecone (>=8.1.0,<9.0.0)",
    "langchain-openai (>=0.3.0,<1.0.0)",
    "langchain-text-splitters (>=1.1.1,<2.0.0)",
    "tabulate (>=0.10.0,<0.11.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "tenacity (>=8.2.0,<10.0.0)",
    "cachetools (>=5.5.0,<6.0.0)",
    "requests (>=2.32.0,<3.0.0)",
    "httpx (>=0.28.0,<1.0.0)",
    "cohere (>=5.0.0,<6.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
